# Tools: make non DEBUG SITL use -march=native