let the compiler optimise a little more SITL when we aren't debugging. 
it also reduces the binary size
No impact on build time.