@peterbarker 2020-09-06
> No impact on build time.

Does it actually make things faster, 'though?

An autotest run might be sufficient to tell us that.  It would be surprising if it made things slower, but you never know....