@khancyr 2020-09-09
Master :
╭─khancyr@pop-os ~/Workspace/ardupilot ‹master*› 
╰─$ build/sitl/benchmarks/benchmark_matrix
2020-09-09 10:11:32
Running build/sitl/benchmarks/benchmark_matrix
Run on (8 X 2300 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 64 KiB (x4)
  L2 Unified 512 KiB (x4)
  L3 Unified 4096 KiB (x1)
Load Average: 1.96, 4.05, 4.36
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
BM_MatrixMultiplication       6.59 ns         6.59 ns     96481023


PR:

╭─khancyr@pop-os ~/Workspace/ardupilot ‹marchnavitve*› 
╰─$ build/sitl/benchmarks/benchmark_matrix
2020-09-09 10:10:27
Running build/sitl/benchmarks/benchmark_matrix
Run on (8 X 2300 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 64 KiB (x4)
  L2 Unified 512 KiB (x4)
  L3 Unified 4096 KiB (x1)
Load Average: 1.62, 4.51, 4.52
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
BM_MatrixMultiplication       5.11 ns         5.10 ns    125113950


So little gain in speed !